Add login, logout and update pages
The login page is used to redirect the user to Strava.

Logout deletes the auth token and redirects to the main page, which
itself will redirect to the login page, assuming the deletion worked.

The Update page will show the status of the modifications being made to
Strava. It currently does nothing.

Signed-off-by: Joel Stanley <joel@jms.id.au>